edit schema

+++ db/schema.sql
@@ -9,9 +9,9 @@ CREATE TABLE snacks (
     name TEXT NOT NULL,
     image TEXT,
     calories INT,
-    fiber INT,
-    sodium INT,
-    sugar INT,
+    fiber DECIMAL,
+    sodium DECIMAL,
+    sugar DECIMAL,
     gluten_free BOOLEAN,
     flavor_profile TEXT,
     is_healthy BOOLEAN
